Move test to module using fresh db for each test to avoid session flush
